Add test for relationship comparison transitivity
This test reveals a problem in RelationshipComparator. It creates 3
relationships A, B and C. Then it confirms that A < B and B < C. In the
end it tests that A < C, which is the transitivity rule. It shows that in
this case with given relationships it is not true.

The relationships are intentionally chosen to be of the same type creating
dependency chain, i.e. A's target is B's declaring and B's target is C's
declaring. The comparator tries to sort consequential relationships correctly,
but if you compare relationships A and C, you are missing the info that there
is a relationship B between them and they are evaluated as equal from the
sorting point of view, which violates the general contract. There is no way
to get this info in comparator from the given relationships, so we have to
drop this sorting feature and instead compare/sort them in some safer way.

The same comparison approach is used in PathComparator, but it is not that
easy to write a test for it.