Fix NotFoundExceptions and NullPointerExceptions
After fixing of the relationship deserialization in #63 on commandline
there were some NotFoundExceptions similar to those in #55 followed by
some NullPointerExceptions somewhere in neo4j. Unfortunately Gnome
collapsed and the console got closed and the exceptions are not in indy log,
so the stacktraces are lost.

Anyway during #63 debugging I noticed that TypeAndClassifier instances
are not detached in Conversions.convertToDetachedRelationships(), so the
produced SimpleArtifactRef instances in target field contained
NeoTypeAndClassifier instances referencing RelationshipProxy. This is
obviously wrong.